proc: Don't change mount options on remount failure.
Introduce proc_mount_opts and use that to ensure proc_parse_options
only works on a temporary copy of the mount options.  This ensures
that on failure the actual proc mount options are not changed.

While less serious this fixes another bug that was introduced
with the introduction of proc mount options.

Fixes: 97412950b10e ("procfs: parse mount options")
Cc: stable@vger.kernel.org
Signed-off-by: "Eric W. Biederman" <ebiederm@xmission.com>