btrfs: fix missing superblock update in the device delete commit transaction
When a device is deleted, the btrfs_super_block::number_devices is
reduced by 1, but we do that after the committing transaction, so this
change may not make it to the disk and waits for the next commit
transaction whenever it happens.

This can be easily demonstrated using the following test case where I
use the btrfs device ready command to read the disk and report.

  mkfs.btrfs -fq -dsingle -msingle $dev1 $dev2
  mount $dev1 /btrfs
  btrfs dev del $dev2 /btrfs
  btrfs dev ready $dev1; echo RESULT=$? <-- 1
  # Without this patch RESULT returns 1, indicating not ready!

Testing with a seed device:

  mkfs.btrfs -fq $dev1
  btrfstune -S1 $dev1
  mount $dev1 /btrfs
  btrfs dev add -f $dev2 /btrfs
  umount /btrfs
  mount $dev2 /btrfs
  btrfs dev del $dev1 /btrfs
  btrfs dev ready $dev2; echo RESULT=$? <-- 1

Fix this by bringing in the num_devices update under the current
transaction.  Also align the local variable declarations in the
function btrfs_rm_dev_item. Delete the todo comment about transient
inconsistent state that's now fixed.

Signed-off-by: Anand Jain <anand.jain@oracle.com>
Reviewed-by: David Sterba <dsterba@suse.com>
Signed-off-by: David Sterba <dsterba@suse.com>