ARM: dts: logicpd-som-lv: Fix OMAP35/DM37 Compatibility issues
The SOM-LV has two models, OMAP35 and AM/DM37.  Most of the pins
are the same, however there are a few monior tweaks that need to
happen where the OMAP35 and DM37 differ.  This causes some crashes
because the range and/or offsets are slightly different between
OMAP35 and DM37.

This patch moves the offending differences from the common file
and places their respective changes into the corresponding dts files.

Fixes: 1e5540b61ec8 ("ARM: dts: Add minimal support for LogicPD
OMAP35xx Torpedo devkit")

Signed-off-by: Adam Ford <aford173@gmail.com>
Signed-off-by: Tony Lindgren <tony@atomide.com>