fsinfo: Add an attribute that lists all the visible mounts in a namespace
Add a filesystem attribute that exports a list of all the visible mounts in
a namespace, given the caller's chroot setting.  The returned list is an
array of:

	struct fsinfo_mount_child {
		__u64	mnt_unique_id;
		__u32	mnt_id;
		__u32	parent_id;
		__u32	mnt_notify_sum;
		__u32	sb_notify_sum;
	};

where each element contains a once-in-a-system-lifetime unique ID, the
mount ID (which may get reused), the parent mount ID and sums of the
notification/change counters for the mount and its superblock.

This works with a read lock on the namespace_sem, but ideally would do it
under the RCU read lock only.

Signed-off-by: David Howells <dhowells@redhat.com>